Corrige problema no envio de dados da ficha dermatológica e updateDermatologica
