--- front/src/components/Fichas/SolicitacaoDeExameAninhar/index.module.css
@@ -1,135 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@300;500;600;700&display=swap");
-
-.container {
-  display: block;
-  width: 100%;
-  max-width: 65rem;
-  min-width: 300px;
-  height: auto;
-  margin: auto;
-  margin-top: 30px;
-  margin-bottom: 30px;
-  font-family: "Poppins", sans-serif;
-}
-
-.container h1 {
-  font-size: 24px;
-  font-weight: 700;
-  text-align: center;
-  margin-top: 50px;
-  color: #333;
-}
-.titulo{
-  max-width: 1118px;
-  width: 100%;
-  height: 44px;
-  background-color: #EEF8F3;
-  color: #000000;
-  font-size: 20px;
-  padding: 9.5px;
-  padding-left: 16px;
-  margin-bottom: 18px;
-  border-radius: 10px;
-  font-weight: 500;
-}
-
-.form_box {
-  border-radius: 20px;
-  border: 0.5px solid #b1b1b1;
-  padding: 20px;
-  background-color: #fff;
-}
-
-.row {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  margin-bottom: 15px;
-}
-
-.anamnesecontainer {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
-  gap: 10px;
-  margin-bottom: 15px;
-}
-
-.anamnesecontainer label {
-  display: flex;
-  align-items: center;
-  background-color: #f9f9f9;
-  padding: 10px;
-  border-radius: 8px;
-  border: 1px solid #ccc;
-  font-size: 14px;
-  cursor: pointer;
-  transition: background 0.3s ease-in-out;
-}
-
-.anamnesecontainer input[type="checkbox"] {
-  margin-right: 8px;
-  accent-color: #39805f;
-}
-
-.anamnesecontainer label:hover {
-  background-color: #e3f1ea;
-}
-
-.column {
-  flex: 1;
-  min-width: 250px;
-  margin: 10px;
-}
-
-.column label {
-  display: block;
-  font-size: 14px;
-  font-weight: 600;
-  margin-bottom: 5px;
-}
-
-.column input,
-.column select,
-.column textarea {
-  width: 100%;
-  padding: 10px;
-  font-size: 14px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-}
-
-.checkbox-group {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 10px;
-  margin-bottom: 15px;
-}
-
-.checkbox-group label {
-  font-size: 14px;
-  display: flex;
-  align-items: center;
-}
-
-.checkbox-group input {
-  margin-right: 5px;
-}
-
-.submitButton {
-  width: 100%;
-  max-width: 180px;
-  height: 50px;
-  border-radius: 10px;
-  border: 1px solid #39805f;
-  color: #fff;
-  background-color: #39805f;
-  font-size: 16px;
-  font-weight: 600;
-  display: block;
-  margin: 20px auto;
-  cursor: pointer;
-}
-
-.submitButton:hover {
-  background-color: #2c6649;
-}
